requirements.txt
# Telegram Bot
python-telegram-bot==20.7

# Database
asyncpg==0.29.0
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9

# API
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.3
python-multipart==0.0.6

# Utils
python-dotenv==1.0.0
redis==5.0.1
httpx==0.25.2
pillow==10.1.0